Generalize AbstractReviewQuestionsField
